chore: refactoring link preview
